backend/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
filelock==3.16.1
fsspec==2024.12.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
setuptools==75.6.0
sqlparse==0.5.3
sympy==1.13.1
torch==2.5.1
torchaudio==2.5.1
typing_extensions==4.12.2
